Change 'Learn lua in 15 minutes' link to 'Learn Lua in Y Minutes'
'Learn lua in 15 minutes' seems to be an older link to the authors blog. The page on learnxinyminutes.com was actually submitted by Tyler Neylon (the creator of 'learn lua in 15 minutes') himself, and considering that the website is more popular, and open source, it will likely be more up-to-date.